requirements.txt
playwright==1.56.0
Jinja2==3.1.6
PyYAML==6.0.3
